+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 - (i18n) Added Danish translation of "Rule"
 - (i18n) Added Dutch translation of "Rule"
 - (i18n) Added Esperanto translation of "Rule"
+- [Ruby] Added `Gherkin::Query#parent_locations` for determining a scenario's parents' line numbers ([#89](https://github.com/cucumber/gherkin/pull/89))
 
 ## [26.2.0] - 2023-04-07
 ### Changed